+++ frontend/components/QuoteForm.js
@@ -1,4 +1,5 @@
-import React, { useReducer } from 'react'
+import React, { useReducer, useContext} from 'react'
+import { DummyContext } from '../context/quotesContext'
 
 const CHANGE_INPUT = 'CHANGE_INPUT'
 const RESET_FORM = 'RESET_FORM'
@@ -21,8 +22,9 @@ const reducer = (state, action) => {
   }
 }
 
-export default function TodoForm(props) {
-  const { createQuote } = props
+export default function TodoForm() {
+  const providedData = useContext(DummyContext)
+  const { createQuote } = providedData
   const [state, dispatch] = useReducer(reducer, initialState)
   const onChange = ({ target: { name, value } }) => {
     dispatch({ type: CHANGE_INPUT, payload: { name, value } })
