+++ frontend/index.js
@@ -3,10 +3,13 @@ import './styles/styles.css'
 import React from 'react'
 import { createRoot } from 'react-dom/client'
 import App from './components/App'
+import {QuotesProvider} from "./context/quotesContext"
 
 const domNode = document.getElementById('root')
 const root = createRoot(domNode)
 
 root.render(
-  <App />
+  <QuotesProvider>
+    <App />
+  </QuotesProvider>
 )
